Remove stdint from api

+++ api/api_types.h
@@ -1,17 +1,14 @@
 #ifndef API_TYPES_H
 #define API_TYPES_H
 
-#include <stdint.h>
-#include <stdbool.h>
-
 // SM API Types
 // --------------------
 
-typedef uint64_t phys_ptr_t;
+typedef unsigned long long phys_ptr_t;
 typedef phys_ptr_t enclave_id_t;
 typedef phys_ptr_t thread_id_t;
-typedef uint64_t region_id_t;
-typedef uint64_t mailbox_id_t;
+typedef unsigned long long region_id_t;
+typedef unsigned long long  mailbox_id_t;
 
 // reserved, "special" enclave_id values. These must not be valid (this implementation requires enclave_id_t be page-aligned, so these values are safe).
 #define OWNER_UNTRUSTED (1)
